+++ gems/aws-sdk-resourceexplorer2/lib/aws-sdk-resourceexplorer2/endpoints.rb
@@ -19,6 +19,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -32,6 +33,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -45,6 +47,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -58,6 +61,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -71,6 +75,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -84,6 +89,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -97,6 +103,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -110,6 +117,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -123,6 +131,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -136,6 +145,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -149,6 +159,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -162,6 +173,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -175,6 +187,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -188,6 +201,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -201,6 +215,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -214,6 +229,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -227,6 +243,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -240,6 +257,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -253,6 +271,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -266,6 +285,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )
@@ -279,6 +299,7 @@ module Aws::ResourceExplorer2
         end
         Aws::ResourceExplorer2::EndpointParameters.new(
           region: context.config.region,
+          use_dual_stack: context.config.use_dualstack_endpoint,
           use_fips: context.config.use_fips_endpoint,
           endpoint: endpoint,
         )